Working on the 'IsLitByLight' function; somethings funky here.

+++ source/ObjectLoader.cpp
@@ -119,16 +119,16 @@ HRESULT ObjectLoader::LoadObject(ID3D11DeviceContext* p_deviceContext, char* p_o
 
 			CustomPrimitiveStruct::TriangleDescription l_triangleDesc;
 			
-			l_triangleDesc.Point1 = point_index1-1;
-			l_triangleDesc.Point2 = point_index2-1;
-			l_triangleDesc.Point3 = point_index3-1;
+			l_triangleDesc.Point1 = static_cast<float>(point_index1 - 1);
+			l_triangleDesc.Point2 = static_cast<float>(point_index2 - 1);
+			l_triangleDesc.Point3 = static_cast<float>(point_index3 - 1);
 			
-			l_triangleDesc.TexCoord1 = texCoord_index1-1;
-			l_triangleDesc.TexCoord2 = texCoord_index2-1;
-			l_triangleDesc.TexCoord3 = texCoord_index3-1;
+			l_triangleDesc.TexCoord1 = static_cast<float>(texCoord_index1 - 1);
+			l_triangleDesc.TexCoord2 = static_cast<float>(texCoord_index2 - 1);
+			l_triangleDesc.TexCoord3 = static_cast<float>(texCoord_index3 - 1);
 			l_triangleDesc.PADDING1 = 1.0f;
 			
-			l_triangleDesc.NormalIndex = normal_index1 - 1;
+			l_triangleDesc.NormalIndex = static_cast<float>(normal_index1 - 1);
 
 			float ambient = 0.001f;
 			float diffuse = 1.0f;
